fix(gateway): split admission mode contract
